Break signature stripper out into its own transformer
