+++ src/main/java/net/minecraftforge/fart/internal/RenamingTransformer.java
@@ -19,21 +19,13 @@
 
 package net.minecraftforge.fart.internal;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Iterator;
-import java.util.Locale;
-import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Consumer;
-import java.util.jar.Attributes;
-import java.util.jar.Manifest;
 import java.util.stream.Collectors;
 
 import org.objectweb.asm.ClassReader;
@@ -45,15 +37,12 @@ import net.minecraftforge.fart.api.Transformer;
 import net.minecraftforge.srgutils.IMappingFile;
 
 public class RenamingTransformer implements Transformer {
-    private static final Attributes.Name SHA_256_DIGEST = new Attributes.Name("SHA-256-Digest");
     private static final String ABSTRACT_FILE = "fernflower_abstract_parameter_names.txt";
     private final EnhancedRemapper remapper;
     private final Set<String> abstractParams = ConcurrentHashMap.newKeySet();
-    private final Consumer<String> log;
 
     public RenamingTransformer(Inheritance inh, IMappingFile map, Consumer<String> log) {
         this.remapper = new EnhancedRemapper(inh, map, log);
-        this.log = log;
     }
 
     @Override
@@ -73,54 +62,10 @@ public class RenamingTransformer implements Transformer {
     }
 
     @Override
-    public ManifestEntry process(ManifestEntry entry) {
-        // Remove all signature entries
-        // see signed jar spec: https://docs.oracle.com/javase/7/docs/technotes/guides/jar/jar.html#Signed_JAR_File
-        try {
-            final Manifest manifest = new Manifest(new ByteArrayInputStream(entry.getData()));
-            boolean found = false;
-            for (final Iterator<Map.Entry<String, Attributes>> it = manifest.getEntries().entrySet().iterator(); it.hasNext();) {
-                final Map.Entry<String, Attributes> section = it.next();
-                for (final Iterator<Map.Entry<Object, Object>> attrIter = section.getValue().entrySet().iterator(); attrIter.hasNext();) {
-                    final Map.Entry<Object, Object> attribute = attrIter.next();
-                    final String key = attribute.getKey().toString().toLowerCase(Locale.ROOT); // spec says this is case-insensitive
-                    if (key.endsWith("-digest")) { // assume that this is a signature entry
-                        attrIter.remove();
-                        found = true;
-                    }
-                    // keep going even if we've found an attribute -- multiple hash formats can be specified for each file
-                }
-
-                if (section.getValue().isEmpty()) {
-                    it.remove();
-                }
-            }
-            if (found) {
-                try (final ByteArrayOutputStream os = new ByteArrayOutputStream()) {
-                    manifest.write(os);
-                    return ManifestEntry.create(entry.getTime(), os.toByteArray());
-                }
-            }
-        } catch (final IOException ex) {
-            log.accept("Failed to remove signature entries from manifest: " + ex);
-        }
-        return entry;
-    }
-
-    @Override
     public ResourceEntry process(ResourceEntry entry) {
         if (ABSTRACT_FILE.equals(entry.getName()))
             return null;
 
-        // Signature metadata
-        if (entry.getName().startsWith("META-INF/")) {
-            if (entry.getName().endsWith(".RSA")
-                    || entry.getName().endsWith(".SF")
-                    || entry.getName().endsWith(".DSA")
-                    || entry.getName().endsWith(".EC")) { // supported by InstallerRewriter but not referenced in the spec
-                return null;
-            }
-        }
         return entry;
     }
